news_summarizer/requirements.txt
streamlit
langchain
langchain-google-genai
requests
beautifulsoup4
python-dotenv
